fix: move event based dependency converter to dependencies file
